// src/app/interfaces/iaboutme.ts
export interface Iaboutme {
	id: String;
	bannerImage: String;
	profilePicture: String;
	ubication: String;
	institution: String;
	institutionImage: String;
	posicion: String;
	descripcion: String;
	bannerImage2: String;
	profilePicture2: String;
	about: String;
	company: [name: String, url: String, logo: String];
}